First almost functional Graph Builder
graph_builder.c now reads .csv files in the required format and stores nodes and edges in a graph structure that is written to a binary file called Graph.bin. An attempt to read that binary file from Astar.c has been made but is failing at the present state. The graph structure and typing need to be reviewed as well as cleaning the code. :)