Working binary write/read
Code still needs to be tidied up, the data structure nodes is writen and read succesfully from graph_builder.c to Astar.c through the binary file. A major bug was found, when assigning the tokenized name strings in the variable pch to the atributes node[i].name or node[i].successors[j].wayname somehow they point to pch instead of storing the value at the time the assignation is made.... And then when recalling this values the only thing you get is the last fetched value for pch which is the first relation line in the csv.